integration_test/requirements.txt
grpcio==1.60.0
protobuf==4.25.1
pytest==7.4.0
